Move get_entries_query to MetatadaStore
